refactor(gpu): extract shared blend-modes WGSL into gpu-shared
Move BLEND_MODES_WGSL out of gpu-compositor so gpu-media can depend on
it without reaching laterally into another lib. Both libs now import
from the new gpu-shared module, eliminating the cross-lib coupling.
