+++ src/lib/gpu-media/media-blend-pipeline.ts
@@ -1,6 +1,6 @@
 import type { BlendMode } from '@/types/blend-modes'
 import { BLEND_MODE_INDEX } from '@/types/blend-modes'
-import { BLEND_MODES_WGSL } from '@/lib/gpu-compositor/blend-modes'
+import { BLEND_MODES_WGSL } from '../gpu-shared/blend-modes'
 
 const MEDIA_BLEND_SHADER = /* wgsl */ `
 struct VertexOutput {
